move environment file
