--- shared/constants/flask/index.js
@@ -1,1 +1,0 @@
-export * from './environment';
